requirements.txt
python-dotenv
requests
django-redis
